Корректирует Entity в соответствие с таблицами БД:
- вносит изменения в типы полей
- привязывает к конкретным таблицам и столбцам;
- устанавливаем связи один-ко-многим, многие-к-одному;
- отключает мапперы, так как требуется их доработка;
- делает сервисы заглушки, для того чтобы запустить приложение.

+++ src/main/java/ru/skypro/homework/repository/AdsRepository.java
@@ -9,6 +9,6 @@ import ru.skypro.homework.entity.AdEntity;
  */
 
 @Repository
-public interface AdsRepository extends JpaRepository<AdEntity, Long> {
+public interface AdsRepository extends JpaRepository<AdEntity, Integer> {
 
 }
